requirements.txt
django                       
djangorestframework
djangorestframework-simplejwt
dj-rest-auth
django-allauth
python-dotenv
requests

